Use lower levels of Docker.DotNet more directly (#20)
This merges in portions of the Docker.DotNet code that deal with the lowest layers of communication (named pipes, proxies, and the like) so we can use them more directly. Adjusted it to suit our needs better, such as by using observables instead of plain streams and switching to System.Text.Json. No longer using Docker.DotNet's wrappers for the REST calls, largely to make error handling easier but also to simplify constructing input/output models.